refactor: refactored prover
