+++ operator/bin/uc_fixture.rs
@@ -5,10 +5,12 @@ use ibc_core_client_types::Height as IbcHeight;
 use ibc_core_commitment_types::commitment::CommitmentRoot;
 use ibc_core_host_types::identifiers::ChainId;
 use serde::{Deserialize, Serialize};
-use sp1_ics07_tendermint_operator::SP1ICS07TendermintProgram;
 use sp1_ics07_tendermint_operator::{
-    rpc::TendermintRPCClient, SP1ICS07TendermintProver, UpdateClientProgram,
-    VerifyMembershipProgram,
+    prover::{
+        SP1ICS07TendermintProgram, SP1ICS07TendermintProver, UpdateClientProgram,
+        VerifyMembershipProgram,
+    },
+    rpc::TendermintRPCClient,
 };
 use sp1_ics07_tendermint_shared::types::sp1_ics07_tendermint::{
     ClientState, Height, TrustThreshold,
